Extract Socket Exceptions into separate classes. Messages - add socketOpeningErrorMessage
