+++ src/main/java/echoserver/SocketCloseException.java
@@ -1,0 +1,8 @@
+package echoserver;
+
+class SocketCloseException extends RuntimeException {
+
+    public SocketCloseException(Exception cause) {
+        super(cause);
+    }
+}
