+++ src/main/java/echoserver/SocketOpenException.java
@@ -1,0 +1,8 @@
+package echoserver;
+
+class SocketOpenException extends RuntimeException {
+
+    public SocketOpenException(Exception cause) {
+        super(cause);
+    }
+}
